@@ -1,6 +1,6 @@
 {
   "name": "@accordproject/concerto-core",
-  "version": "3.5.0",
+  "version": "3.6.0",
   "description": "Core Implementation for the Concerto Modeling Language",
   "homepage": "https://github.com/accordproject/concerto",
   "engines": {
@@ -65,15 +65,15 @@
     "sinon-chai": "3.7.0",
     "tmp-promise": "3.0.2",
     "typescript": "4.6.3",
-    "webpack": "5.64.2",
+    "webpack": "5.76.0",
     "webpack-cli": "4.9.1",
     "xregexp": "5.1.1",
     "yargs": "17.3.1"
   },
   "dependencies": {
-    "@accordproject/concerto-cto": "3.5.0",
-    "@accordproject/concerto-metamodel": "3.5.0",
-    "@accordproject/concerto-util": "3.5.0",
+    "@accordproject/concerto-cto": "3.6.0",
+    "@accordproject/concerto-metamodel": "3.6.0",
+    "@accordproject/concerto-util": "3.6.0",
     "dayjs": "1.10.8",
     "debug": "4.3.1",
     "lorem-ipsum": "2.0.3",
