@@ -16,9 +16,8 @@
 
 const { MetaModelNamespace } = require('@accordproject/concerto-metamodel');
 
-const Decorated = require('./decorated');
+const Declaration = require('./declaration');
 const IllegalModelException = require('./illegalmodelexception');
-const ModelUtil = require('../modelutil');
 const NumberValidator = require('./numbervalidator');
 const StringValidator = require('./stringvalidator');
 
@@ -26,8 +25,8 @@ const StringValidator = require('./stringvalidator');
 /* eslint-disable no-unused-vars */
 /* istanbul ignore next */
 if (global === undefined) {
-    const ModelFile = require('./modelfile');
     const Validator = require('./validator');
+    const ClassDeclaration = require('./classdeclaration');
 }
 /* eslint-enable no-unused-vars */
 
@@ -42,31 +41,7 @@ if (global === undefined) {
  * @class
  * @memberof module:concerto-core
  */
-class ScalarDeclaration extends Decorated {
-    /**
-     * Create a ScalarDeclaration from an Abstract Syntax Tree. The AST is the
-     * result of parsing.
-     *
-     * @param {ModelFile} modelFile - the ModelFile for this class
-     * @param {Object} ast - the AST created by the parser
-     * @throws {IllegalModelException}
-     */
-    constructor(modelFile, ast) {
-        super(ast);
-        this.modelFile = modelFile;
-        this.process();
-    }
-
-    /**
-     * Returns the ModelFile that defines this class.
-     *
-     * @public
-     * @return {ModelFile} the owning ModelFile
-     */
-    getModelFile() {
-        return this.modelFile;
-    }
-
+class ScalarDeclaration extends Declaration {
     /**
      * Process the AST and build the model
      *
@@ -76,7 +51,6 @@ class ScalarDeclaration extends Decorated {
     process() {
         super.process();
 
-        this.name = this.ast.name;
         this.superType = null;
         this.superTypeDeclaration = null;
         this.idField = null;
@@ -120,8 +94,6 @@ class ScalarDeclaration extends Decorated {
         } else {
             this.defaultValue = null;
         }
-
-        this.fqn = ModelUtil.getFullyQualifiedName(this.modelFile.getNamespace(), this.name);
     }
 
     /**
@@ -152,36 +124,9 @@ class ScalarDeclaration extends Decorated {
     }
 
     /**
-     * Returns the short name of a class. This name does not include the
-     * namespace from the owning ModelFile.
-     *
-     * @return {string} the short name of this class
-     */
-    getName() {
-        return this.name;
-    }
-
-    /**
-     * Return the namespace of this class.
-     * @return {string} namespace - a namespace.
-     */
-    getNamespace() {
-        return this.modelFile.getNamespace();
-    }
-
-    /**
-     * Returns the fully qualified name of this class.
-     * The name will include the namespace if present.
-     *
-     * @return {string} the fully-qualified name of this class
-     */
-    getFullyQualifiedName() {
-        return this.fqn;
-    }
-
-    /**
      * Returns false as scalars are never identified.
      * @returns {Boolean} false as scalars are never identified
+     * @deprecated
      */
     isIdentified() {
         return false;
@@ -190,6 +135,7 @@ class ScalarDeclaration extends Decorated {
     /**
      * Returns false as scalars are never identified.
      * @returns {Boolean} false as scalars are never identified
+     * @deprecated
      */
     isSystemIdentified() {
         return false;
@@ -197,12 +143,12 @@ class ScalarDeclaration extends Decorated {
 
     /**
      * Returns null as scalars are never identified.
-     * @return {null} null, as scalars are never identified
+     * @return {string} as scalars are never identified
+     * @deprecated
      */
     getIdentifierFieldName() {
         return null;
     }
-
 
     /**
      * Returns the FQN of the super type for this class or null if this
@@ -215,15 +161,20 @@ class ScalarDeclaration extends Decorated {
     }
 
     /**
-     * Throws an error as scalars do not have supertypes.
+     * Returns the FQN of the super type for this class or null if this
+     * class does not have a super type.
+     *
+     * @return {string} the FQN name of the super type or null
+     * @deprecated
      */
     getSuperType() {
-        throw new Error('Scalars do not have super types.');
+        return null;
     }
 
     /**
      * Get the super type class declaration for this class.
-     * @return {ScalarDeclaration | null} the super type declaration, or null if there is no super type.
+     * @return {ClassDeclaration} the super type declaration, or null if there is no super type.
+     * @deprecated
      */
     getSuperTypeDeclaration() {
         return null;
@@ -238,9 +189,9 @@ class ScalarDeclaration extends Decorated {
     }
 
     /**
-       * Returns the default value for the field or null
-       * @return {string | number | null} the default value for the field or null
-       */
+     * Returns the default value for the field or null
+     * @return {string | number | null} the default value for the field or null
+     */
     getDefaultValue() {
         if(this.defaultValue) {
             return this.defaultValue;
@@ -262,72 +213,10 @@ class ScalarDeclaration extends Decorated {
      * Returns true if this class is abstract.
      *
      * @return {boolean} true if the class is abstract
+     * @deprecated
      */
     isAbstract() {
         return true;
-    }
-
-    /**
-     * Returns true if this class is the definition of an asset.
-     *
-     * @return {boolean} true if the class is an asset
-     */
-    isAsset() {
-        return false;
-    }
-
-    /**
-     * Returns true if this class is the definition of a participant.
-     *
-     * @return {boolean} true if the class is a participant
-     */
-    isParticipant() {
-        return false;
-    }
-
-    /**
-     * Returns true if this class is the definition of a transaction.
-     *
-     * @return {boolean} true if the class is a transaction
-     */
-    isTransaction() {
-        return false;
-    }
-
-    /**
-     * Returns true if this class is the definition of an event.
-     *
-     * @return {boolean} true if the class is an event
-     */
-    isEvent() {
-        return false;
-    }
-
-    /**
-     * Returns true if this class is the definition of a concept.
-     *
-     * @return {boolean} true if the class is a concept
-     */
-    isConcept() {
-        return false;
-    }
-
-    /**
-     * Returns true if this class is the definition of an enum.
-     *
-     * @return {boolean} true if the class is an enum
-     */
-    isEnum() {
-        return false;
-    }
-
-    /**
-     * Returns true if this class is the definition of a class declaration.
-     *
-     * @return {boolean} true if the class is a class
-     */
-    isClassDeclaration() {
-        return false;
     }
 
     /**
@@ -338,6 +227,57 @@ class ScalarDeclaration extends Decorated {
     isScalarDeclaration() {
         return true;
     }
+
+    /**
+     * Returns true if this class is the definition of an asset.
+     *
+     * @return {boolean} true if the class is an asset
+     * @deprecated
+     */
+    isAsset() {
+        return false;
+    }
+
+    /**
+     * Returns true if this class is the definition of a participant.
+     *
+     * @return {boolean} true if the class is a participant
+     * @deprecated
+     */
+    isParticipant() {
+        return false;
+    }
+
+    /**
+     * Returns true if this class is the definition of a transaction.
+     *
+     * @return {boolean} true if the class is a transaction
+     * @deprecated
+     */
+    isTransaction() {
+        return false;
+    }
+
+    /**
+     * Returns true if this class is the definition of an event.
+     *
+     * @return {boolean} true if the class is an event
+     * @deprecated
+     */
+    isEvent() {
+        return false;
+    }
+
+    /**
+     * Returns true if this class is the definition of a concept.
+     *
+     * @return {boolean} true if the class is a concept
+     * @deprecated
+     */
+    isConcept() {
+        return false;
+    }
+
 }
 
 module.exports = ScalarDeclaration;
