@@ -14,7 +14,8 @@
 
 'use strict';
 
-const DiagramVisitor = require('../common/diagramvisitor');
+const { ModelUtil } = require('@accordproject/concerto-core');
+const DiagramVisitor = require('../../../common/diagramvisitor');
 
 /**
  * Convert the contents of a ModelManager
@@ -22,7 +23,7 @@ const DiagramVisitor = require('../common/diagramvisitor');
  * Set a fileWriter property (instance of FileWriter) on the parameters
  * object to control where the generated code is written to disk.
  *
- * @private
+ * @protected
  * @class
  * @memberof module:concerto-tools
  */
@@ -31,7 +32,7 @@ class PlantUMLVisitor extends DiagramVisitor {
      * Visitor design pattern
      * @param {ModelManager} modelManager - the object being visited
      * @param {Object} parameters  - the parameter
-     * @private
+     * @protected
      */
     visitModelManager(modelManager, parameters) {
         parameters.fileWriter.openFile('model.puml');
@@ -50,7 +51,7 @@ class PlantUMLVisitor extends DiagramVisitor {
      * Visitor design pattern
      * @param {ClassDeclaration} classDeclaration - the object being visited
      * @param {Object} parameters  - the parameter
-     * @private
+     * @protected
      */
     visitAssetDeclaration(classDeclaration, parameters) {
         this.writeDeclaration(classDeclaration, parameters, '(A,green)' );
@@ -60,7 +61,7 @@ class PlantUMLVisitor extends DiagramVisitor {
      * Visitor design pattern
      * @param {ClassDeclaration} classDeclaration - the object being visited
      * @param {Object} parameters  - the parameter
-     * @private
+     * @protected
      */
     visitEnumDeclaration(classDeclaration, parameters) {
         this.writeDeclaration(classDeclaration, parameters, '(E,grey)' );
@@ -70,7 +71,7 @@ class PlantUMLVisitor extends DiagramVisitor {
      * Visitor design pattern
      * @param {ClassDeclaration} classDeclaration - the object being visited
      * @param {Object} parameters  - the parameter
-     * @private
+     * @protected
      */
     visitParticipantDeclaration(classDeclaration, parameters) {
         this.writeDeclaration(classDeclaration, parameters, '(P,lightblue)' );
@@ -80,7 +81,7 @@ class PlantUMLVisitor extends DiagramVisitor {
      * Visitor design pattern
      * @param {ClassDeclaration} classDeclaration - the object being visited
      * @param {Object} parameters  - the parameter
-     * @private
+     * @protected
      */
     visitTransactionDeclaration(classDeclaration, parameters) {
         this.writeDeclaration(classDeclaration, parameters, '(T,yellow)' );
@@ -91,7 +92,7 @@ class PlantUMLVisitor extends DiagramVisitor {
      * Visitor design pattern
      * @param {ClassDeclaration} classDeclaration - the object being visited
      * @param {Object} parameters  - the parameter
-     * @private
+     * @protected
      */
     visitClassDeclaration(classDeclaration, parameters) {
         this.writeDeclaration(classDeclaration, parameters );
@@ -102,7 +103,7 @@ class PlantUMLVisitor extends DiagramVisitor {
      * @param {ClassDeclaration} classDeclaration - the object being visited
      * @param {Object} parameters  - the parameter
      * @param {string} [style] - the style for the prototype (optional)
-     * @private
+     * @protected
      */
     writeDeclaration(classDeclaration, parameters, style) {
         parameters.fileWriter.writeLine(0, 'class ' + this.escapeString(classDeclaration.getFullyQualifiedName()) + (style ? ` << ${style} >> ` : ' ') + '{' );
@@ -113,8 +114,8 @@ class PlantUMLVisitor extends DiagramVisitor {
 
         if (!classDeclaration.isEnum()){
             classDeclaration.getOwnProperties().forEach((property) => {
-                const source = classDeclaration.getFullyQualifiedName();
-                const target = property.getFullyQualifiedTypeName();
+                const source = this.escapeString(classDeclaration.getFullyQualifiedName());
+                const target = this.escapeString(property.getFullyQualifiedTypeName());
                 const label = property.getName();
                 const array = property.isArray() ? '"*"':'"1"';
                 if (property.isRelationship?.()) {
@@ -131,13 +132,29 @@ class PlantUMLVisitor extends DiagramVisitor {
     }
 
     /**
-     * Escape versions and periods.
-     * @param {String} string - the object being visited
+     * Escape fully qualified names. We preserve the dots in the
+     * package name, remove the '@' symbol because it is invalid
+     * and remove the dots in the version (because otherwise packages get created)
+     * @param {String} input - the object being visited
      * @return {String} string  - the parameter
-     * @private
+     * @protected
      */
-    escapeString(string) {
-        return string.replace('@', '_');
+    escapeString(input) {
+        const hasNamespace = ModelUtil.getNamespace(input) !== '';
+        if(hasNamespace) {
+            const typeName = ModelUtil.getShortName(input);
+            const ns = ModelUtil.getNamespace(input);
+            const {name,version} = ModelUtil.parseNamespace(ns);
+            if(version) {
+                return `${name}_${version.replace(/\./g, '_')}.${typeName}`;
+            }
+            else {
+                return `${name}.${typeName}`;
+            }
+        }
+        else {
+            return input;
+        }
     }
 }
 
