@@ -14,7 +14,8 @@ export var CodeGen: {
     ProtobufVisitor: typeof import("./lib/codegen/fromcto/protobuf/protobufvisitor");
     OpenApiVisitor: typeof import("./lib/codegen/fromcto/openapi/openapivisitor");
     AvroVisitor: typeof import("./lib/codegen/fromcto/avro/avrovisitor");
-    InferFromJsonSchema: typeof import("./lib/codegen/fromJsonSchema/cto/inferModel");
+    JSONSchemaToConcertoVisitor: typeof import("./lib/codegen/fromJsonSchema/cto/jsonSchemaVisitor");
+    OpenApiToConcertoVisitor: typeof import("./lib/codegen/fromOpenApi/cto/openApiVisitor");
     formats: {
         golang: typeof import("./lib/codegen/fromcto/golang/golangvisitor");
         jsonschema: typeof import("./lib/codegen/fromcto/jsonschema/jsonschemavisitor");
@@ -32,4 +33,5 @@ export var CodeGen: {
         avro: typeof import("./lib/codegen/fromcto/avro/avrovisitor");
     };
 };
+export var Common: typeof import("./lib/common/common");
 export var version: any;
